+++ src/app/app.module.ts
@@ -3,8 +3,8 @@ import {BrowserModule} from '@angular/platform-browser';
 
 import {AppComponent} from './app.component';
 import {BreweryListComponent} from './components/search-components/brewery-list/brewery-list.component';
-import {NavbarComponent} from './components/shared-components/navbar/navbar.component';
-import {FooterComponent} from './components/shared-components/footer/footer.component';
+import {NavbarComponent} from './components/core-components/navbar/navbar.component';
+import {FooterComponent} from './components/core-components/footer/footer.component';
 import {StoreModule} from "@ngrx/store";
 import {metaReducers, reducers} from "./store";
 import {ReactiveFormsModule} from "@angular/forms";
@@ -21,14 +21,14 @@ import {MatInputModule} from "@angular/material/input";
 import {MatIconModule} from "@angular/material/icon";
 import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
 import {BreweryDetailComponent} from './components/search-components/brewery-detail/brewery-detail.component';
-import {DatePipe} from "./pipes/date-pipe";
+import {DatePipe} from "./shared/pipes/date-pipe";
 import {SearchBarComponent} from "./components/search-components/search-bar/search-bar.component";
 import {SearchHistoryComponent} from './components/search-components/search-history/search-history.component';
 import {MatListModule} from "@angular/material/list";
 import {CommonModule} from "@angular/common";
 import {SearchContainerComponent} from './components/search-components/search-container/search-container.component';
-import {LoaderComponent} from './components/shared-components/loader/loader.component';
-import {InfoBoxComponent} from './components/shared-components/info-box/info-box.component';
+import {LoaderComponent} from './components/core-components/loader/loader.component';
+import {InfoBoxComponent} from './shared/components/info-box/info-box.component';
 
 @NgModule({
   declarations: [
