+++ src/app/components/core-components/footer/footer.component.spec.ts
