+++ src/app/components/core-components/footer/footer.component.ts
