+++ src/app/components/core-components/loader/loader.component.scss
