+++ src/app/components/core-components/loader/loader.component.ts
