+++ src/app/components/core-components/navbar/navbar.component.scss
