+++ src/app/components/search-components/search-history/search-history.component.spec.ts
@@ -6,7 +6,7 @@ import {selectSearchHistory} from "../../../store/selectors/selectors";
 import {initialApplicationState, SearchHistory} from "../../../store/models/model";
 import {By} from "@angular/platform-browser";
 import {CUSTOM_ELEMENTS_SCHEMA, DebugElement} from "@angular/core";
-import {DatePipe} from "../../../pipes/date-pipe";
+import {DatePipe} from "../../../shared/pipes/date-pipe";
 import {setSearchHistory} from "../../../store/actions/actions";
 import {MatIconModule} from "@angular/material/icon";
 
@@ -21,7 +21,7 @@ describe('SearchHistoryComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports:[ MatIconModule],
+      imports: [MatIconModule],
       declarations: [SearchHistoryComponent, DatePipe],
       providers: [provideMockStore({initialState: initialApplicationState})],
       schemas: [CUSTOM_ELEMENTS_SCHEMA]
